+++ client/src/components/App.jsx
@@ -1,8 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import axios from 'axios';
 import { Switch, Route } from 'react-router-dom';
-import { Launcher } from 'react-chat-window';
-import Container from '@mui/material/Container';
+
 import Navigation from './Navigation.jsx';
 import ProductsPage from './Pages/ProductsPage.jsx';
 import BoxPage from './Pages/BoxPage.jsx';
@@ -11,6 +10,7 @@ import CartPage from './Pages/CartPage.jsx';
 import AccountPage from './Pages/AccountPage.jsx';
 import LoginPage from './Pages/LoginPage/LoginPage.jsx';
 import ProductViewPage from './Pages/ProductViewPage.jsx';
+import Chat from './Chat/Chat.jsx';
 // import useAuth from '../context/AuthContext.jsx';
 import useMainContext from '../context/MainContext.jsx';
 import useAuth from '../context/AuthContext.jsx';
@@ -20,39 +20,7 @@ import Footer from './Footer.jsx';
 export default function App() {
   const { page } = useMainContext();
   const { currentUser } = useAuth();
-  const [messageList, setMessageList] = useState([]);
-
-  function sendMessage(text) {
-    if (text.length > 0) {
-      setMessageList([
-        ...messageList,
-        {
-          author: 'them',
-          type: 'text',
-          data: { text },
-        },
-      ]);
-    }
-  }
-  function onMessageWasSent(message) {
-    const reply = {
-      author: 'Nutritionist',
-      type: 'text',
-      data: { text: "We have received your reply, you'll hear back soon." },
-    };
-
-    setMessageList([...messageList, message, reply]);
-  }
-  useEffect(() => {
-    const message = {
-      author: 'Nutritionist',
-      type: 'text',
-      data: {
-        text: "Welcome to the chat, please leave any questions for a nutritionist and we'll get back to you shortly.",
-      },
-    };
-    setMessageList([...messageList, message]);
-  }, []);
+  console.log(currentUser);
 
   // test endpoint and server connection
   useEffect(() => {
@@ -96,21 +64,13 @@ export default function App() {
         <Navigation />
       </nav>
       <section className="content">{renderPage()}</section>
-      <section>
-        <Container sx={{ 'div.sc-launcher': { zIndex: 100 } }}>
-          <Launcher
-            agentProfile={{
-              teamName: 'Chat with a Nutritionist!',
-              imageUrl:
-                'https://a.slack-edge.com/66f9/img/avatars-teams/ava_0001-34.png',
-            }}
-            onMessageWasSent={onMessageWasSent}
-            messageList={messageList}
-            showEmoji
-            mute
-          />
-        </Container>
-      </section>
+      {currentUser ? (
+        <section>
+          <Chat />
+        </section>
+      ) : (
+        ''
+      )}
       <Footer />
     </>
   );
